move `split_stream` to `s3`
